--- src/binary_cookies_parser/__init__.py
@@ -1,4 +1,0 @@
-from binary_cookies_parser._deserialize import load, loads
-from binary_cookies_parser._serialize import dump, dumps
-
-__all__ = ["dump", "dumps", "load", "loads"]
