+++ src/binarycookies/parser.py
@@ -2,8 +2,8 @@ from typing import List
 
 from typing_extensions import deprecated
 
-from binary_cookies_parser._deserialize import load
-from binary_cookies_parser.models import Cookie
+from binarycookies._deserialize import load
+from binarycookies.models import Cookie
 
 
 @deprecated("read_binary_cookies_file is deprecated, use binary_cookies_parser.load instead.")
